+++ packages/kokoas-client/src/api/andpad/convertProjToAndpad.test.ts
@@ -10,5 +10,7 @@ describe('convertProjToAndpad', () => {
     const result = await convertProjToAndpad('8aa54bb3-10e7-45fa-adde-df7776082c77');
 
     console.log(result);
+
+    expect(result).toBeDefined();
   });
 });
