+++ packages/kokoas-client/src/pages/projRegister/sections/RecordSelect/SaveToAndpadDialog.tsx
@@ -2,7 +2,7 @@ import { Button, Dialog, DialogActions, DialogTitle } from '@mui/material';
 import { AndpadButton } from 'kokoas-client/src/components/ui/buttons/AndpadButton';
 import { AndpadLogo } from 'kokoas-client/src/components/ui/icons';
 import { useURLParams } from 'kokoas-client/src/hooks/useURLParams';
-import { useSaveAndpadProject } from 'kokoas-client/src/hooksQuery';
+import { useConvertToAndpadByProjId, useSaveAndpadProject } from 'kokoas-client/src/hooksQuery';
 import { TypeOfForm } from '../../form';
 import { SaveToAndpadDialogContent } from './SaveToAndpadDialogContent';
 
@@ -14,14 +14,17 @@ export const SaveToAndpadDialog = ({
   handleClose: () => void
 }) => {
   const { projId } = useURLParams<TypeOfForm>();
-  const { mutate: mutateAndpad } = useSaveAndpadProject(projId || '');
+  const { data, isLoading } = useConvertToAndpadByProjId(open ? projId : '');
+
+  const { mutate: mutateAndpad } = useSaveAndpadProject();
 
 
   const handleClick = () => {
-    mutateAndpad();
+    if (data) {
+      mutateAndpad(data);
+    }
     handleClose();
   };
-
 
   return (
     <Dialog
@@ -29,11 +32,12 @@ export const SaveToAndpadDialog = ({
       onClose={handleClose}
       keepMounted={false}
       maxWidth={'xs'}
+      fullWidth
     >
       <DialogTitle>
         アンドパッドへ登録しますか。
       </DialogTitle>
-      <SaveToAndpadDialogContent />
+      <SaveToAndpadDialogContent isLoading={isLoading} convertedData={data} />
       <DialogActions>
         <Button onClick={handleClose}>
           キャンセル
@@ -41,6 +45,7 @@ export const SaveToAndpadDialog = ({
         <AndpadButton
           onClick={handleClick}
           startIcon={<AndpadLogo />}
+          disabled={isLoading}
         >
           はい
         </AndpadButton>
