+++ packages/kokoas-client/src/pages/projRegister/sections/RecordSelect/SaveToAndpadDialogContent.tsx
@@ -1,12 +1,24 @@
 import { DialogContent, DialogContentText } from '@mui/material';
+import { SaveProjectData } from 'api-andpad';
+import { Loading } from 'kokoas-client/src/components/ui/loading/Loading';
+import { SaveToAndpadDialogData } from './SaveToAndpadDialogData';
 
-export const SaveToAndpadDialogContent = () => {
+export const SaveToAndpadDialogContent = ({
+  convertedData,
+  isLoading,
+
+}:{
+  isLoading: boolean
+  convertedData?: SaveProjectData
+}) => {
 
   return (
-    <DialogContent>
+    <DialogContent sx={{ height: '90vh' }}>
       <DialogContentText>
-        Andpadへ格納します。
+        {isLoading ?  '少々お待ちください。' : '以下の内容でAndpadへ保存されます。'}
       </DialogContentText>
+      {isLoading && <Loading /> }
+      {!!convertedData && <SaveToAndpadDialogData convertedData={convertedData} />}
     </DialogContent>
   );
 };
